# chore(tests): Remove coveragePathIgnorePatterns from fxa-payments-server