--- a/packages/fxa-payments-server/server/jest.config.js
+++ b/packages/fxa-payments-server/server/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
-  coveragePathIgnorePatterns: ['<rootDir>'],
-  collectCoverageFrom: ['**/*.js', '!**/jest*js'],
+  // coveragePathIgnorePatterns: ['<rootDir>'],
+  collectCoverageFrom: ['**/*.js', '!bin/*', '!coverage/**', '!**/jest*js'],
   // TO DO: update this file once more server tests are in place
   coverageThreshold: {
     global: {